requirements.txt
# A requirements file for use with the pip package manager
Django<1.6
python-twitter
